# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ScreenTinker Public API
  version: 1.9.29
  description: |
    Public, token-scoped REST API for ScreenTinker digital signage.

    ## Authentication
    Every endpoint (except the two unauthenticated device-render endpoints) accepts a
    **scoped personal access token** presented as `Authorization: Bearer st_...`. The
    dashboard's session JWT is also accepted, but this contract targets token consumers.

    ## Scope ladder (read &#60; write &#60; full)
    Tokens carry one of three scopes; higher scopes subsume lower ones:
    - `read`  — all `GET` endpoints.
    - `write` — all resource mutations (`POST` / `PUT` / `PATCH` / `DELETE`).
    - `full`  — operational fleet commands (currently only
      `POST /groups/{id}/command`: reboot / shutdown / screen on/off).

    Each operation advertises its minimum scope via the `x-required-scope` extension
    and restates it in the operation description. A token with insufficient scope is
    rejected with `403`.

    All resources are scoped to the token's workspace; cross-workspace access is not
    available to tokens.
servers:
  - url: /api
...
security:
  - ApiToken: []
  - SessionJWT: []
components:
  securitySchemes:
    ApiToken:
      type: http
      scheme: bearer
      bearerFormat: "st_<token>"
      description: |
        Scoped personal access token. Send as `Authorization: Bearer st_...`.
        The token's scope (`read`, `write`, or `full`) gates which operations
        it may call (see the scope ladder in the API description).
    SessionJWT:
      type: http
      scheme: bearer
      description: |
        Dashboard session JWT. Also accepted on every authenticated endpoint,
        but this spec targets scoped-token consumers.
...
paths:
...
  /walls:
    get:
      tags: [walls]
      summary: List video walls (with attached devices)
      description: 'Requires scope: read.'
      x-required-scope: read
      responses:
        '200': { description: Array of walls. }
    post:
      tags: [walls]
      summary: Create a video wall
      description: 'Requires scope: write.'
      x-required-scope: write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                grid_cols: { type: integer }
                grid_rows: { type: integer }
                bezel_h_mm: { type: number }
                bezel_v_mm: { type: number }
                playlist_id: { type: string }
      responses:
        '201': { description: Created wall. }
  /walls/{id}:
    get:
      tags: [walls]
      summary: Get a video wall with devices
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Wall with devices. }
    put:
      tags: [walls]
      summary: Update a video wall
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                grid_cols: { type: integer }
                grid_rows: { type: integer }
                bezel_h_mm: { type: number }
                bezel_v_mm: { type: number }
                sync_mode: { type: string }
                leader_device_id: { type: string }
                content_id: { type: string }
                playlist_id: { type: string }
      responses:
        '200': { description: Updated wall. }
    delete:
      tags: [walls]
      summary: Delete a video wall
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: '{ success }.' }
  /walls/{id}/devices:
    put:
      tags: [walls]
      summary: Set the wall's device grid positions (replaces member set)
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [devices]
              properties:
                devices:
                  type: array
                  items:
                    type: object
                    properties:
                      device_id: { type: string }
                      grid_col: { type: integer }
                      grid_row: { type: integer }
                      rotation:
                        type: integer
                        enum: [0, 90, 180, 270]
                        default: 0
                        description: >-
                          How this panel is physically mounted, as degrees CLOCKWISE that its image
                          must be turned to come out upright on the wall (the same convention as a
                          device's `orientation`). canvas_* are in WALL space — the wall as the
                          audience sees it — so a portrait-mounted 1920x1080 panel is a tall tile
                          with rotation 90, and content needs no pre-rotating. Anything other than
                          0/90/180/270 is stored as 0. While a panel is in a wall this replaces its
                          own `orientation`, so the two can never rotate the content twice.
                      canvas_x: { type: number }
                      canvas_y: { type: number }
                      canvas_width: { type: number }
                      canvas_height: { type: number }
      responses:
        '200': { description: Updated wall with devices. }
  /walls/{id}/content:
    put:
      tags: [walls]
      summary: Set wall content (legacy single-video path)
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content_id: { type: string }
      responses:
        '200': { description: '{ success }.' }
  /walls/{id}/device-config/{deviceId}:
    get:
      tags: [walls]
      summary: Get wall config for a specific member device
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Per-device wall config. }

  # ------------------------------------------------------------------ reports
...